requirements.txt
argparse
configparser==4.0.2
python-daemon
paho-mqtt
smbus2
bme280
pylint